Refactor rating hash logic into reusable hook
Moved hash parsing logic for initial rating selection into a new useRatingHash hook in src/libs/utils/rating.ts. Updated dialog-new-rating and all-ratings components to use this hook, improving code reuse and maintainability.
